# feat(architecture): implement 5-layer backend architecture with pure business logic [DAN-1]