--- a/apps/__init__.py
+++ b/apps/__init__.py
@@ -1,0 +1,1 @@
+# Apps package