… .claude/hooks/api-standards-checker.py
… .claude/hooks/commit-message-validator.py
… .claude/hooks/import-organizer.py
… .claude/hooks/no-reflexive-agreement.py
… .claude/hooks/notification.py
… .claude/hooks/pnpm-enforcer.py
… .claude/hooks/post_tool_use.py
… .claude/hooks/pre_compact.py
… .claude/hooks/pre_tool_use.py
… .claude/hooks/session_start.py
… .claude/hooks/stop.py
… .claude/hooks/subagent_stop.py
… .claude/hooks/task-completion-enforcer.py
… .claude/hooks/typescript-validator.py
… .claude/hooks/universal-linter.py
… .claude/hooks/user_prompt_sumbit.py
… .claude/hooks/utils/llm/oai.py
… .claude/hooks/utils/tts/elevenlabs_tts.py
… .claude/hooks/utils/tts/pyttsx3_tts.py
… .claude/settings.json
… .claude/status-lines/status_line.py
… .claude/status-lines/status_line_v2.py
… .claude/status-lines/status_line_v3.py
… .claude/status-lines/status_line_v4.py
… .code/agents/1/exec-call_xW6BiHVLeKJNd0ZtSaAXiafu.txt
… .code/agents/18/exec-call_OqT33rheGazdhVRNLm1MzTQM.txt
… .code/agents/20/exec-call_1tLiM7Sbc954dIUmGZ52Sclh.txt
… .code/agents/20/exec-call_NRQZjYNGdSmEUtHuY2boB7y1.txt
… .code/agents/44ce4209-f081-4345-b2bf-916aa437d889/result.txt
… .code/agents/4e6f6437-8cc0-428a-9327-efab4862d789/result.txt